Image retrieval improved for online search - store to temp instead of keeping byte[] in memory

+++ Modules/datacrow-core/src/main/java/org/datacrow/core/services/SearchTask.java
@@ -37,7 +37,6 @@ import java.util.Map;
 import javax.imageio.ImageIO;
 
 import org.datacrow.core.DcConfig;
-import org.datacrow.core.http.HttpConnectionUtil;
 import org.datacrow.core.log.DcLogManager;
 import org.datacrow.core.log.DcLogger;
 import org.datacrow.core.objects.DcImageIcon;
@@ -384,43 +383,29 @@ public abstract class SearchTask extends Thread {
     }
     
     protected DcImageIcon getImage(String url) {
-    	
     	url = url.replace("http://", "https://");
     	
         try {
 			Image img = ImageIO.read(new URL(url));
 			DcImageIcon icon = new DcImageIcon(img);
 			
-			// write to temp folder
-			File file = new File(CoreUtilities.getTempFolder(), CoreUtilities.getUniqueID() + ".jpg");
-			CoreUtilities.writeMaxImageToFile(icon, file);
-			
-			// delete the file on exit of Data Crow
-			file.deleteOnExit();
-			
-			// flush the in memory image
-			icon.flush();
-			
-			// send image pointing to local disk storage
-			return new DcImageIcon(file);
+			if (icon.getIconHeight() > 50) {
+    			// write to temp folder
+    			File file = new File(CoreUtilities.getTempFolder(), CoreUtilities.getUniqueID() + ".jpg");
+    			CoreUtilities.writeMaxImageToFile(icon, file);
+    			
+    			// delete the file on exit of Data Crow
+    			file.deleteOnExit();
+    			
+    			// flush the in memory image
+    			icon.flush();
+    			
+    			// send image pointing to local disk storage
+    			return new DcImageIcon(file);
+			}
         } catch (Exception e) {
             logger.debug("Cannot download image from [" + url + "]", e);
         }
         return null;    	
-    }
-    
-    @Deprecated
-    protected byte[] getImageBytes(String url) {
-        url = url.replace("http://", "https://");
-        try {
-            if (url != null && url.length() > 0) {
-                byte[] b = HttpConnectionUtil.retrieveBytes(url);
-                if (b != null && b.length > 50)
-                    return b;
-            }
-        } catch (Exception e) {
-            logger.debug("Cannot download image from [" + url + "]", e);
-        }
-        return null;
     }
 }
